stacking, remove stacking connector
Remove the stacking connector to give us more space to expand the
display ribbon cable slot. On the Pi5, there can be up to two ribbon
cables which are located slightly closer to the corner of the board. The
previous configuration was hard to wire.

Extend the ribbon cable slot to the edge of the board.

The board can still be stacked - control signals will still pass through
the Raspberry Pi headers, and the end user will just need to wire the
power input to all stacked boards separately.