v2.4, add estop/global disable
Add an input to allow all outputs to be controlled remotely. Inputs to
MOSFETs are adapted to incorporate an AND gate, which has the nice
advantage of buffering the output so the LEDs will be brighter.

Clamp inputs to the AND gates to 5.1V with a Zener to prevent damage.

ERC and DRC passing.